Rename t3c plugin_verifier to t3c-verify
