+++ cache-config/t3c-verify/test-files/etc/astats.config
