+++ cache-config/t3c-verify/test-files/etc/bad-plugin.config
