+++ cache-config/t3c-verify/test-files/etc/bad-remap-multiline.config
