+++ cache-config/t3c-verify/test-files/etc/hdr_rw_mid_bar.config
