+++ cache-config/t3c-verify/test-files/etc/plugin.config
