+++ cache-config/t3c-verify/test-files/etc/regex_revalidate.config
