+++ cache-config/t3c-verify/test-files/libexec/plugin.so
